MDL-38086 Make .label.modtype_label -> .mod_label
Changes the CSS classname used on lists of resources/activities
on course pages from .label and .modtype_label to simply using
the more commen naming convention of .mod_label.

Label is just an example, all resource and activity types are affected
by this change.

I've also updated the few places in core where the old style
was referred to in CSS or JS.