MDL-41840 Restore: Page title, breadcrumbs show incorrect stage
Unfortunately it isn't possible to display the current stage before
printing header any more, because we need to be able to show a
progress bar while doing the code that considers whether or not
priunt stage. So this change removes stage information from page
title. It is still present (in bold) in the strip that shows where
you are within the restore/import.

This display was inconsistent anyway; change makes it consistent
between Restore and Import.