MDL-44486 progress bar: move styles to theme
Removes the styles that were attached via style tags and replaced
them with class names to pick up existing progress bar styles used
in Bootstrap. This also worked in Base derived themes since someone
has hardcoded a progress bar using those classes and backported
the CSS already.

(Note, maybe that usage could call this function now? I also noticed
that the Lesson mod has it's own progress bar thing happening too)

This generally looks much better, except for when the wibbler is
used, but the wibbler is rather ugly looking anyway, so there's
probably a call to fix that in a seperate bug rather than hold this
one up.

Oh, this change also removes the hardcoded color change from
amber to green as I'm not sure it was necessary. If needed it could
be done with either warning/success classes, or entirely with styles
based on width[100%] I think.